requirements.txt
flask
gunicorn
pymongo[srv]
python-dotenv
python-binance
pandas
ta
skyfield
